Add folder for resources and split views
