+++ portailva/file/views/__init__.py
@@ -1,0 +1,4 @@
+from .admin_resource_files import *
+from .admin_files import *
+from .association_files_management import *
+from .files import *
